requirements.txt
flask
openai-whisper
imageio-ffmpeg
anthropic
